+++ game_scripts/factories/rooms/keys_doors_puzzle_factory.lua
@@ -73,6 +73,7 @@ local GOAL_ID = -1
 local PARAMS_WHITELIST = {
     episodeLengthSeconds = true,
     invocationMode = true,
+    allowHoldOutLevels = true,
     levelGenerator = true,
     playerId = true,
     players = true,