+++ game_scripts/levels/tests/seed_test.lua
@@ -28,7 +28,7 @@ function tests.testSeed()
   local seed = 2
 
   local api = require('levels.' .. testLevelScript:gsub('/', '.'))
-  api:init{invocationMode = "testbed", datasetPath = 'dummy'}
+  api:init{allowHoldOutLevels = 'true', datasetPath = 'dummy'}
   api:start(0, seed)
   local rand1 = random:uniformReal(0, 1)
   local rand2 = random:uniformReal(0, 1)